consolidate model runners
